Fix argument type checking
During the update to support python 3 a bug was introduced into the
argument type checking logic that resulted in the argument value being
set to None when a valid type check occured.

Additionally this commit fixes the tests to account for the fact that
the argument dictionary can be mutated by the checking logic. In this
case this was not expected and resulted in tests passing that should
have been failing.